.history/backend/requirements_20241125180054.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.1
timm==0.9.12
python-jose==3.3.0
python-dotenv==1.0.0